refactor: use v2r 모듈 3개 통합
